chore(*): Moving mrl specific logic in its own service and setting navigation service runtime if column layouts are defined.
