+++ projects/igniteui-angular/src/lib/grids/grid/grid.component.ts
@@ -30,6 +30,7 @@ import { IgxGridSelectionService, IgxGridCRUDService } from '../../core/grid-sel
 import { IgxOverlayService } from '../../services/index';
 import { IgxForOfSyncService } from '../../directives/for-of/for_of.sync.service';
 import { IgxDragIndicatorIconDirective } from '../row-drag.directive';
+import { IgxGridMRLNavigationService } from '../grid-mrl-navigation.service';
 
 let NEXT_ID = 0;
 
@@ -855,6 +856,7 @@ export class IgxGridComponent extends IgxGridBaseComponent implements IGridDataB
         if (this.hideGroupedColumns && this.columnList && this.groupingExpressions) {
             this._setGroupColsVisibility(this.hideGroupedColumns);
         }
+        this._setupNavigationService();
     }
 
     public ngOnInit() {
@@ -899,6 +901,13 @@ export class IgxGridComponent extends IgxGridBaseComponent implements IGridDataB
             return this.extractDataFromSelection(source);
         } else {
             return super.getSelectedData();
+        }
+    }
+
+    private _setupNavigationService() {
+        if (this.hasColumnLayouts) {
+            this.navigation = new IgxGridMRLNavigationService();
+            this.navigation.grid = this;
         }
     }
 }
